+++ cognite/experimental/data_classes/contextualization.py
@@ -23,6 +23,10 @@ class EntityMatchingMatchRule(CogniteResource):
         extractors=None,
         priority=None,
         matches=None,
+        flags=None,
+        conflicts=None,
+        overlaps=None,
+        number_of_matches=None,
         num_conflicts=None,
         num_overlaps=None,
         cognite_client=None,
@@ -31,6 +35,10 @@ class EntityMatchingMatchRule(CogniteResource):
         self.extractors = extractors
         self.priority = priority
         self.matches = matches
+        self.flags = flags
+        self.conflicts = conflicts
+        self.overlaps = overlaps
+        self.number_of_matches = number_of_matches
         self.num_conflicts = num_conflicts
         self.num_overlaps = num_overlaps
         self._cognite_client = cognite_client
@@ -38,12 +46,13 @@ class EntityMatchingMatchRule(CogniteResource):
     def _repr_html_(self):
         extractors = _label_groups(copy.deepcopy(self.extractors), self.conditions)
 
-        info_df = super().to_pandas(camel_case=True).drop(["extractors", "conditions"])
+        info_df = super().to_pandas(camel_case=True)
         info_df = dataframe_summarize(info_df)
         info_df.index.name = "Stats"
-
-        coloured_matches_table = _color_matches(extractors, self.matches)
-        return info_df._repr_html_() + coloured_matches_table
+        html = info_df._repr_html_()
+        if self.matches is not None:
+            html += _color_matches(extractors, self.matches)
+        return html
 
 
 class EntityMatchingMatchRuleList(CogniteResourceList):
@@ -144,12 +153,17 @@ class EntityMatchingPipelineRun(ContextualizationJob):
         return self.result["suggestedRules"]
 
     @property
+    def errors(self):
+        """Returns list of error messages encountered while running. Depends on .result and may block"""
+        return self.get("errors")
+
+    @property
     def generated_rules(self):
         """List of suggested new match rules. Depends on .result and may block"""
         return EntityMatchingMatchRuleList._load(self.result["generatedRules"], cognite_client=self._cognite_client)
 
     @property
-    def matches(self) -> EntityMatchingMatchRuleList:
+    def matches(self) -> EntityMatchingMatchList:
         """List of matches. Depends on .result and may block"""
         matches = self.result["matches"]
         match_fields = (self.pipeline.model_parameters or {}).get("matchFields", [{"source": "name", "target": "name"}])
@@ -331,6 +345,43 @@ class EntityMatchingPipelineUpdate(CogniteUpdate):  # not implemented yet
 class EntityMatchingPipelineList(CogniteResourceList):
     _RESOURCE = EntityMatchingPipeline
     _UPDATE = EntityMatchingPipelineUpdate
+
+
+class MatchRulesSuggestJob(ContextualizationJob):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._status_path = "/context/matchrules/suggest/"
+
+    @property
+    def rules(self) -> EntityMatchingMatchRuleList:
+        """Depends on .result and may block"""
+        return EntityMatchingMatchRuleList._load(self.result["rules"])
+
+    def _repr_html_(self):
+        rules = self.rules  # TODO: optional loading? before super() for status
+        df = super().to_pandas()
+        df.loc["rules"] = [f"{len(rules)} items"]
+        return df._repr_html_()
+
+
+class MatchRulesApplyJob(ContextualizationJob):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._status_path = "/context/matchrules/suggest/"
+
+    @property
+    def rules(self) -> EntityMatchingMatchRuleList:
+        """Depends on .result and may block"""
+        reformated_rules = [{**item["rule"], **item} for item in self.result["items"]]
+        for rule in reformated_rules:
+            rule.pop("rule")
+        return EntityMatchingMatchRuleList._load(reformated_rules)
+
+    def _repr_html_(self):
+        rules = self.rules  # TODO: optional loading? before super() for status
+        df = super().to_pandas()
+        df.loc["rules"] = [f"{len(rules)} items"]
+        return df._repr_html_()
 
 
 class PNIDDetection(CogniteResource):